refactor: better handling for registry urls (#21545)
