+++ cli/args/mod.rs
@@ -104,6 +104,64 @@ pub fn npm_registry_default_url() -> &'static Url {
   &NPM_REGISTRY_DEFAULT_URL
 }
 
+pub fn deno_registry_url() -> &'static Url {
+  static DENO_REGISTRY_URL: Lazy<Url> = Lazy::new(|| {
+    let env_var_name = "DENO_REGISTRY_URL";
+    if let Ok(registry_url) = std::env::var(env_var_name) {
+      // ensure there is a trailing slash for the directory
+      let registry_url = format!("{}/", registry_url.trim_end_matches('/'));
+      match Url::parse(&registry_url) {
+        Ok(url) => {
+          return url;
+        }
+        Err(err) => {
+          log::debug!(
+            "Invalid {} environment variable: {:#}",
+            env_var_name,
+            err,
+          );
+        }
+      }
+    }
+
+    deno_graph::source::DEFAULT_DENO_REGISTRY_URL.clone()
+  });
+
+  &DENO_REGISTRY_URL
+}
+
+pub fn deno_registry_api_url() -> &'static Url {
+  static DENO_REGISTRY_API_URL: Lazy<Url> = Lazy::new(|| {
+    let env_var_name = "DENO_REGISTRY_API_URL";
+    if let Ok(registry_url) = std::env::var(env_var_name) {
+      // ensure there is a trailing slash for the directory
+      let registry_url = format!("{}/", registry_url.trim_end_matches('/'));
+      match Url::parse(&registry_url) {
+        Ok(url) => {
+          return url;
+        }
+        Err(err) => {
+          log::debug!(
+            "Invalid {} environment variable: {:#}",
+            env_var_name,
+            err,
+          );
+        }
+      }
+    }
+
+    let host = deno_graph::source::DEFAULT_DENO_REGISTRY_URL
+      .host_str()
+      .unwrap();
+
+    let mut url = deno_graph::source::DEFAULT_DENO_REGISTRY_URL.clone();
+    url.set_host(Some(&format!("api.{}", host))).unwrap();
+    url
+  });
+
+  &DENO_REGISTRY_API_URL
+}
+
 pub fn ts_config_to_emit_options(
   config: deno_config::TsConfig,
 ) -> deno_ast::EmitOptions {
@@ -1854,5 +1912,13 @@ mod test {
         temp_dir_path.join("nested/foo/bazz.ts"),
       ]
     )
+  }
+
+  #[test]
+  fn deno_registry_urls() {
+    let reg_url = deno_registry_url();
+    assert!(reg_url.as_str().ends_with('/'));
+    let reg_api_url = deno_registry_api_url();
+    assert!(reg_api_url.as_str().ends_with('/'));
   }
 }
